Feat: Add Migrations for Postgres database
I used db-migrate to create a migration for my databases so I do not
have to manually set up the database in docker. I can automatically
create tables, etc with the db-migrate up command.

I also made edits to the database.ts file in order to accommodate
testing. I would not want tests to be run on the live database.

I also installed some packages for the database configuration.